Move `TreeNodeProps` to tree types file
